refactor: Move utils.logging into utils.common.logging
